Refactor: `motionStakes` client resolver
