+++ src/data/generated.ts
@@ -286,6 +286,7 @@ export type Query = {
   legacyNumberOfRecoveryRoles: Scalars['Int'];
   loggedInUser: LoggedInUser;
   motionCurrentUserVoted: Scalars['Boolean'];
+  motionStakes: MotionStakes;
   motionUserVoteRevealed: MotionVoteReveal;
   motionVoteResults: MotionVoteResults;
   motionVoterReward: Scalars['String'];
@@ -299,7 +300,6 @@ export type Query = {
   recoveryRolesAndApprovalsForSession: Array<UsersAndRecoveryApprovals>;
   recoveryRolesUsers: Array<User>;
   recoverySystemMessagesForSession: Array<SystemMessage>;
-  stakeMotionLimits: StakeLimits;
   subscribedUsers: Array<User>;
   systemInfo: SystemInfo;
   token: Token;
@@ -412,6 +412,14 @@ export type QueryMotionCurrentUserVotedArgs = {
 };
 
 
+export type QueryMotionStakesArgs = {
+  colonyAddress: Scalars['String'];
+  userAddress: Scalars['String'];
+  motionId: Scalars['Int'];
+  rootHash: Scalars['String'];
+};
+
+
 export type QueryMotionUserVoteRevealedArgs = {
   motionId: Scalars['Int'];
   colonyAddress: Scalars['String'];
@@ -475,14 +483,6 @@ export type QueryRecoveryRolesUsersArgs = {
 export type QueryRecoverySystemMessagesForSessionArgs = {
   blockNumber: Scalars['Int'];
   colonyAddress: Scalars['String'];
-};
-
-
-export type QueryStakeMotionLimitsArgs = {
-  colonyAddress: Scalars['String'];
-  userAddress: Scalars['String'];
-  motionId: Scalars['Int'];
-  rootHash: Scalars['String'];
 };
 
 
@@ -832,7 +832,7 @@ export type ProcessedMetaColony = {
   avatarURL?: Maybe<Scalars['String']>;
 };
 
-export type StakeLimits = {
+export type MotionStakes = {
   remainingToFullyStaked: Scalars['String'];
   maxUserStake: Scalars['String'];
   minUserStake: Scalars['String'];
@@ -1525,7 +1525,14 @@ export type RecoveryAllEnteredEventsQueryVariables = Exact<{
 
 export type RecoveryAllEnteredEventsQuery = { recoveryAllEnteredEvents: Array<Pick<ParsedEvent, 'type' | 'name' | 'values' | 'createdAt' | 'emmitedBy' | 'blockNumber' | 'transactionHash'>> };
 
-export type StakeMotionLimitsQueryVariables = Exact<{
+export type LegacyNumberOfRecoveryRolesQueryVariables = Exact<{
+  colonyAddress: Scalars['String'];
+}>;
+
+
+export type LegacyNumberOfRecoveryRolesQuery = Pick<Query, 'legacyNumberOfRecoveryRoles'>;
+
+export type MotionStakesQueryVariables = Exact<{
   colonyAddress: Scalars['String'];
   userAddress: Scalars['String'];
   motionId: Scalars['Int'];
@@ -1533,14 +1540,7 @@ export type StakeMotionLimitsQueryVariables = Exact<{
 }>;
 
 
-export type StakeMotionLimitsQuery = { stakeMotionLimits: Pick<StakeLimits, 'remainingToFullyStaked' | 'maxUserStake' | 'minUserStake'> };
-
-export type LegacyNumberOfRecoveryRolesQueryVariables = Exact<{
-  colonyAddress: Scalars['String'];
-}>;
-
-
-export type LegacyNumberOfRecoveryRolesQuery = Pick<Query, 'legacyNumberOfRecoveryRoles'>;
+export type MotionStakesQuery = { motionStakes: Pick<MotionStakes, 'remainingToFullyStaked' | 'maxUserStake' | 'minUserStake'> };
 
 export type MotionsSystemMessagesQueryVariables = Exact<{
   motionId: Scalars['Int'];
@@ -3634,44 +3634,6 @@ export function useRecoveryAllEnteredEventsLazyQuery(baseOptions?: Apollo.LazyQu
 export type RecoveryAllEnteredEventsQueryHookResult = ReturnType<typeof useRecoveryAllEnteredEventsQuery>;
 export type RecoveryAllEnteredEventsLazyQueryHookResult = ReturnType<typeof useRecoveryAllEnteredEventsLazyQuery>;
 export type RecoveryAllEnteredEventsQueryResult = Apollo.QueryResult<RecoveryAllEnteredEventsQuery, RecoveryAllEnteredEventsQueryVariables>;
-export const StakeMotionLimitsDocument = gql`
-    query StakeMotionLimits($colonyAddress: String!, $userAddress: String!, $motionId: Int!, $rootHash: String!) {
-  stakeMotionLimits(colonyAddress: $colonyAddress, userAddress: $userAddress, motionId: $motionId, rootHash: $rootHash) @client {
-    remainingToFullyStaked
-    maxUserStake
-    minUserStake
-  }
-}
-    `;
-
-/**
- * __useStakeMotionLimitsQuery__
- *
- * To run a query within a React component, call `useStakeMotionLimitsQuery` and pass it any options that fit your needs.
- * When your component renders, `useStakeMotionLimitsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useStakeMotionLimitsQuery({
- *   variables: {
- *      colonyAddress: // value for 'colonyAddress'
- *      userAddress: // value for 'userAddress'
- *      motionId: // value for 'motionId'
- *      rootHash: // value for 'rootHash'
- *   },
- * });
- */
-export function useStakeMotionLimitsQuery(baseOptions?: Apollo.QueryHookOptions<StakeMotionLimitsQuery, StakeMotionLimitsQueryVariables>) {
-        return Apollo.useQuery<StakeMotionLimitsQuery, StakeMotionLimitsQueryVariables>(StakeMotionLimitsDocument, baseOptions);
-      }
-export function useStakeMotionLimitsLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<StakeMotionLimitsQuery, StakeMotionLimitsQueryVariables>) {
-          return Apollo.useLazyQuery<StakeMotionLimitsQuery, StakeMotionLimitsQueryVariables>(StakeMotionLimitsDocument, baseOptions);
-        }
-export type StakeMotionLimitsQueryHookResult = ReturnType<typeof useStakeMotionLimitsQuery>;
-export type StakeMotionLimitsLazyQueryHookResult = ReturnType<typeof useStakeMotionLimitsLazyQuery>;
-export type StakeMotionLimitsQueryResult = Apollo.QueryResult<StakeMotionLimitsQuery, StakeMotionLimitsQueryVariables>;
 export const LegacyNumberOfRecoveryRolesDocument = gql`
     query LegacyNumberOfRecoveryRoles($colonyAddress: String!) {
   legacyNumberOfRecoveryRoles(colonyAddress: $colonyAddress) @client
@@ -3703,6 +3665,44 @@ export function useLegacyNumberOfRecoveryRolesLazyQuery(baseOptions?: Apollo.Laz
 export type LegacyNumberOfRecoveryRolesQueryHookResult = ReturnType<typeof useLegacyNumberOfRecoveryRolesQuery>;
 export type LegacyNumberOfRecoveryRolesLazyQueryHookResult = ReturnType<typeof useLegacyNumberOfRecoveryRolesLazyQuery>;
 export type LegacyNumberOfRecoveryRolesQueryResult = Apollo.QueryResult<LegacyNumberOfRecoveryRolesQuery, LegacyNumberOfRecoveryRolesQueryVariables>;
+export const MotionStakesDocument = gql`
+    query MotionStakes($colonyAddress: String!, $userAddress: String!, $motionId: Int!, $rootHash: String!) {
+  motionStakes(colonyAddress: $colonyAddress, userAddress: $userAddress, motionId: $motionId, rootHash: $rootHash) @client {
+    remainingToFullyStaked
+    maxUserStake
+    minUserStake
+  }
+}
+    `;
+
+/**
+ * __useMotionStakesQuery__
+ *
+ * To run a query within a React component, call `useMotionStakesQuery` and pass it any options that fit your needs.
+ * When your component renders, `useMotionStakesQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useMotionStakesQuery({
+ *   variables: {
+ *      colonyAddress: // value for 'colonyAddress'
+ *      userAddress: // value for 'userAddress'
+ *      motionId: // value for 'motionId'
+ *      rootHash: // value for 'rootHash'
+ *   },
+ * });
+ */
+export function useMotionStakesQuery(baseOptions?: Apollo.QueryHookOptions<MotionStakesQuery, MotionStakesQueryVariables>) {
+        return Apollo.useQuery<MotionStakesQuery, MotionStakesQueryVariables>(MotionStakesDocument, baseOptions);
+      }
+export function useMotionStakesLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<MotionStakesQuery, MotionStakesQueryVariables>) {
+          return Apollo.useLazyQuery<MotionStakesQuery, MotionStakesQueryVariables>(MotionStakesDocument, baseOptions);
+        }
+export type MotionStakesQueryHookResult = ReturnType<typeof useMotionStakesQuery>;
+export type MotionStakesLazyQueryHookResult = ReturnType<typeof useMotionStakesLazyQuery>;
+export type MotionStakesQueryResult = Apollo.QueryResult<MotionStakesQuery, MotionStakesQueryVariables>;
 export const MotionsSystemMessagesDocument = gql`
     query MotionsSystemMessages($motionId: Int!, $colonyAddress: String!) {
   motionsSystemMessages(motionId: $motionId, colonyAddress: $colonyAddress) @client {
